refactor!: Remove sailthru and braze specific modules
Instead, use a more generically named email module that abstracts
whatever provider we're using under the covers:

ecommerce_worker.email.v1.api.*

The public API are four celery tasks and one bounce-check method:
- send_code_assignment_nudge_email
- send_offer_assignment_email
- send_offer_update_email
- send_offer_usage_email
- did_email_bounce
