+++ ecommerce_worker/configuration/base.py
@@ -1,5 +1,5 @@
 # CELERY
-# Default broker URL. See http://celery.readthedocs.org/en/latest/configuration.html#broker-url.
+# Default broker URL. See http://celery.readthedocs.org/en/4.0/userguide/configuration.html#std:setting-broker_url.
 BROKER_URL = 'redis://:celery@127.0.0.1:6379'
 
 # Disable connection pooling. Connections may be severed by load balancers.
@@ -15,14 +15,14 @@ BROKER_HEARTBEAT = 10.0
 BROKER_HEARTBEAT_CHECKRATE = 2
 
 # Backend used to store task results.
-# See http://celery.readthedocs.org/en/latest/configuration.html#celery-result-backend.
+# See http://celery.readthedocs.org/en/4.0/userguide/configuration.html#std:setting-result_backend.
 CELERY_RESULT_BACKEND = None
 
 # A sequence of modules to import when the worker starts.
-# See http://celery.readthedocs.org/en/latest/configuration.html#celery-imports.
+# See http://celery.readthedocs.org/en/4.0/userguide/configuration.html#std:setting-imports.
 CELERY_IMPORTS = (
+    'ecommerce_worker.email.v1.tasks',
     'ecommerce_worker.fulfillment.v1.tasks',
-    'ecommerce_worker.sailthru.v1.tasks',
 )
 
 DEFAULT_PRIORITY_QUEUE = 'ecommerce.default'
@@ -30,7 +30,7 @@ CELERY_DEFAULT_EXCHANGE = 'ecommerce'
 CELERY_DEFAULT_ROUTING_KEY = 'ecommerce'
 CELERY_DEFAULT_QUEUE = DEFAULT_PRIORITY_QUEUE
 # Prevent Celery from removing handlers on the root logger. Allows setting custom logging handlers.
-# See http://celery.readthedocs.org/en/latest/configuration.html#celeryd-hijack-root-logger.
+# See http://celery.readthedocs.org/en/4.0/userguide/configuration.html#std:setting-worker_hijack_root_logger.
 CELERYD_HIJACK_ROOT_LOGGER = False
 
 # Specify allowed serializers that are consistent with Celery 3 defaults
@@ -66,7 +66,7 @@ SITE_OVERRIDES = None
 # .. toggle_name: BRAZE['BRAZE_ENABLE']
 # .. toggle_implementation: PythonConstant
 # .. toggle_default: False
-# .. toggle_description: Toggle for allowing emails to be sent via Braze instead of sailthru
+# .. toggle_description: Toggle for allowing emails to be sent via Braze
 # .. toggle_use_cases: open_edx
 # .. toggle_creation_date: 2021-02-09
 # .. toggle_target_removal_date: None
@@ -88,46 +88,4 @@ BRAZE = {
     # Retry settings for Braze celery tasks
     'BRAZE_RETRY_SECONDS': 3600,
     'BRAZE_RETRY_ATTEMPTS': 6,
-    }
-
-# Settings for Sailthru email marketing integration
-SAILTHRU = {
-    # Set to false to ignore Sailthru events
-    'SAILTHRU_ENABLE': False,
-
-    # Template used when user upgrades to verified
-    'SAILTHRU_UPGRADE_TEMPLATE': None,
-
-    # Template used with user purchases a course
-    'SAILTHRU_PURCHASE_TEMPLATE': None,
-
-    # Template used with user enrolls in a free course
-    'SAILTHRU_ENROLL_TEMPLATE': None,
-
-    # Abandoned cart template
-    'SAILTHRU_ABANDONED_CART_TEMPLATE': None,
-
-    # minutes to delay before abandoned cart message
-    'SAILTHRU_ABANDONED_CART_DELAY': 60,
-
-    # Sailthru key and secret required for integration
-    'SAILTHRU_KEY': 'sailthru key here',
-    'SAILTHRU_SECRET': 'sailthru secret here',
-
-    # Retry settings for Sailthru celery tasks
-    'SAILTHRU_RETRY_SECONDS': 3600,
-    'SAILTHRU_RETRY_ATTEMPTS': 6,
-
-    # ttl for cached course content from Sailthru (in seconds)
-    'SAILTHRU_CACHE_TTL_SECONDS': 3600,
-
-    # dummy price for audit/honor (i.e., if cost = 0)
-    # Note: setting this value to 0 skips Sailthru calls for free transactions
-    'SAILTHRU_MINIMUM_COST': 100,
-
-    # Transactional email template name map
-    'templates': {
-        'assignment_email': 'Offer Assignment Email',
-        'enterprise_portal_email': 'Enterprise Portal Email',
-    }
 }
