+++ ecommerce_worker/email/v1/api.py
@@ -1,0 +1,17 @@
+"""Public API for ecommerce_worker.email.v1"""
+
+# Disable unused imports because these imports are meant for external consumption
+# pylint: disable=unused-import
+
+# Celery tasks
+from .tasks import (
+    send_code_assignment_nudge_email,
+    send_offer_assignment_email,
+    send_offer_update_email,
+    send_offer_usage_email,
+)
+
+# Normal helper methods
+from .utils import (
+    did_email_bounce,
+)
