--- ecommerce_worker/sailthru/v1/constants.py
@@ -1,1 +1,0 @@
-BASE_ENTERPRISE_URL_VAR_NAME='base_enterprise_url'
