+++ requirements/base.txt
@@ -4,30 +4,64 @@
 #
 #    make upgrade
 #
-amqp==2.6.1               # via kombu
-billiard==3.6.3.0         # via celery
-celery==4.4.7             # via -c requirements/constraints.txt, -r requirements/base.in
-certifi==2020.6.20        # via requests
-chardet==3.0.4            # via requests
-django-crum==0.7.8        # via edx-django-utils
-django-waffle==2.0.0      # via edx-django-utils
-django==2.2.17            # via -c requirements/constraints.txt, django-crum, edx-django-utils
-edx-django-utils==3.11.0  # via edx-rest-api-client
-edx-rest-api-client==5.2.1  # via -r requirements/base.in
-idna==2.10                # via requests
-kombu==4.6.11             # via celery
-newrelic==5.22.1.152      # via edx-django-utils
-pbr==5.5.1                # via stevedore
-psutil==5.7.3             # via edx-django-utils
-pyjwt==1.7.1              # via edx-rest-api-client
-pytz==2020.4              # via celery, django
-redis==3.5.3              # via -r requirements/base.in
-requests==2.24.0          # via edx-rest-api-client, sailthru-client, slumber
-sailthru-client==2.3.5    # via -r requirements/base.in
-simplejson==3.17.2        # via sailthru-client
-six==1.15.0               # via -r requirements/base.in
-slumber==0.7.1            # via edx-rest-api-client
-sqlparse==0.4.1           # via django
-stevedore==3.2.2          # via edx-django-utils
-urllib3==1.25.11          # via requests
-vine==1.3.0               # via amqp, celery
+amqp==2.6.1
+    # via kombu
+billiard==3.6.4.0
+    # via celery
+celery==4.4.7
+    # via
+    #   -c requirements/constraints.txt
+    #   -r requirements/base.in
+certifi==2020.12.5
+    # via requests
+chardet==4.0.0
+    # via requests
+django-crum==0.7.9
+    # via edx-django-utils
+django-waffle==2.1.0
+    # via edx-django-utils
+django==2.2.22
+    # via
+    #   -c requirements/constraints.txt
+    #   django-crum
+    #   edx-django-utils
+edx-django-utils==4.0.0
+    # via edx-rest-api-client
+edx-rest-api-client==5.3.0
+    # via -r requirements/base.in
+idna==2.10
+    # via requests
+kombu==4.6.11
+    # via celery
+newrelic==6.2.0.156
+    # via edx-django-utils
+pbr==5.6.0
+    # via stevedore
+psutil==5.8.0
+    # via edx-django-utils
+pyjwt==2.1.0
+    # via edx-rest-api-client
+pytz==2021.1
+    # via
+    #   celery
+    #   django
+redis==3.5.3
+    # via -r requirements/base.in
+requests==2.25.1
+    # via
+    #   edx-rest-api-client
+    #   slumber
+six==1.16.0
+    # via -r requirements/base.in
+slumber==0.7.1
+    # via edx-rest-api-client
+sqlparse==0.4.1
+    # via django
+stevedore==3.3.0
+    # via edx-django-utils
+urllib3==1.26.4
+    # via requests
+vine==1.3.0
+    # via
+    #   amqp
+    #   celery
