+++ requirements/optional.txt
@@ -4,4 +4,5 @@
 #
 #    make upgrade
 #
-newrelic==5.22.1.152      # via -r requirements/optional.in
+newrelic==6.2.0.156
+    # via -r requirements/optional.in
