+++ requirements/pip_tools.txt
@@ -4,8 +4,12 @@
 #
 #    make upgrade
 #
-click==7.1.2              # via pip-tools
-pip-tools==5.5.0          # via -c requirements/constraints.txt, -r requirements/pip_tools.in
+click==7.1.2
+    # via pip-tools
+pip-tools==5.5.0
+    # via
+    #   -c requirements/constraints.txt
+    #   -r requirements/pip_tools.in
 
 # The following packages are considered to be unsafe in a requirements file:
 # pip
