+++ requirements/test.txt
@@ -6,187 +6,195 @@
 #
 amqp==2.6.1
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   kombu
-astroid==2.3.3
+astroid==2.5.6
     # via
-    #   -r test.in
+    #   -r requirements/test.in
     #   pylint
     #   pylint-celery
-attrs==20.2.0
+attrs==21.2.0
     # via pytest
-billiard==3.6.3.0
+billiard==3.6.4.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   celery
 celery==4.4.7
     # via
-    #   -c constraints.txt
-    #   -r base.txt
-certifi==2020.6.20
+    #   -c requirements/constraints.txt
+    #   -r requirements/base.txt
+certifi==2020.12.5
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   requests
-chardet==3.0.4
+chardet==4.0.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   requests
 click-log==0.3.2
     # via edx-lint
 click==7.1.2
     # via
     #   click-log
+    #   code-annotations
     #   edx-lint
-coverage==5.3
+code-annotations==1.1.1
+    # via edx-lint
+coverage==5.5
     # via
-    #   -r test.in
+    #   -r requirements/test.in
     #   pytest-cov
-ddt==1.4.1
-    # via -r test.in
-django-crum==0.7.8
+ddt==1.4.2
+    # via -r requirements/test.in
+django-crum==0.7.9
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-django-utils
-django-waffle==2.0.0
+django-waffle==2.1.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-django-utils
-django==2.2.17
     # via
-    #   -c constraints.txt
-    #   -r base.txt
+    #   -c requirements/constraints.txt
+    #   -r requirements/base.txt
+    #   code-annotations
     #   django-crum
     #   edx-django-utils
-edx-django-utils==3.11.0
+    #   edx-lint
+edx-django-utils==4.0.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-rest-api-client
-edx-lint==1.5.2
-    # via -r test.in
-edx-rest-api-client==5.2.1
-    # via -r base.txt
+edx-lint==5.0.0
+    # via -r requirements/test.in
+edx-rest-api-client==5.3.0
+    # via -r requirements/base.txt
 idna==2.10
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   requests
 iniconfig==1.1.1
     # via pytest
-isort==4.3.21
+isort==5.8.0
     # via pylint
+jinja2==2.11.3
+    # via code-annotations
 kombu==4.6.11
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   celery
-lazy-object-proxy==1.4.3
+lazy-object-proxy==1.6.0
     # via astroid
+markupsafe==1.1.1
+    # via jinja2
 mccabe==0.6.1
     # via pylint
-mock==4.0.2
-    # via -r test.in
-newrelic==5.22.1.152
+mock==4.0.3
+    # via -r requirements/test.in
+newrelic==6.2.0.156
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-django-utils
-packaging==20.4
+packaging==20.9
     # via pytest
-pbr==5.5.1
+pbr==5.6.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   stevedore
 pluggy==0.13.1
     # via pytest
-psutil==5.7.3
+psutil==5.8.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-django-utils
 py==1.10.0
     # via pytest
-pycodestyle==2.6.0
-    # via -r test.in
-pyjwt==1.7.1
+pycodestyle==2.7.0
+    # via -r requirements/test.in
+pyjwt==2.1.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-rest-api-client
 pylint-celery==0.3
     # via edx-lint
-pylint-django==2.0.11
+pylint-django==2.4.4
     # via
-    #   -r test.in
+    #   -r requirements/test.in
     #   edx-lint
 pylint-plugin-utils==0.6
     # via
-    #   -r test.in
+    #   -r requirements/test.in
     #   pylint-celery
     #   pylint-django
-pylint==2.4.4
+pylint==2.8.2
     # via
-    #   -r test.in
+    #   -r requirements/test.in
     #   edx-lint
     #   pylint-celery
     #   pylint-django
     #   pylint-plugin-utils
 pyparsing==2.4.7
     # via packaging
-pytest-cov==2.10.1
-    # via -r test.in
-pytest==6.1.2
+pytest-cov==2.11.1
+    # via -r requirements/test.in
+pytest==6.2.4
     # via
-    #   -r test.in
+    #   -r requirements/test.in
     #   pytest-cov
-pytz==2020.4
+python-slugify==5.0.2
+    # via code-annotations
+pytz==2021.1
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   celery
     #   django
+pyyaml==5.4.1
+    # via code-annotations
 redis==3.5.3
-    # via -r base.txt
-requests==2.24.0
+    # via -r requirements/base.txt
+requests==2.25.1
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-rest-api-client
     #   responses
-    #   sailthru-client
     #   slumber
-responses==0.12.0
-    # via -r test.in
-sailthru-client==2.3.5
-    # via -r base.txt
-simplejson==3.17.2
+responses==0.13.3
+    # via -r requirements/test.in
+six==1.16.0
     # via
-    #   -r base.txt
-    #   sailthru-client
-six==1.15.0
-    # via
-    #   -r base.txt
-    #   astroid
+    #   -r requirements/base.txt
     #   edx-lint
-    #   packaging
     #   responses
 slumber==0.7.1
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   edx-rest-api-client
 sqlparse==0.4.1
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   django
-stevedore==3.2.2
+stevedore==3.3.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
+    #   code-annotations
     #   edx-django-utils
-testfixtures==6.15.0
-    # via -r test.in
+testfixtures==6.17.1
+    # via -r requirements/test.in
+text-unidecode==1.3
+    # via python-slugify
 toml==0.10.2
-    # via pytest
-urllib3==1.25.11
     # via
-    #   -r base.txt
+    #   pylint
+    #   pytest
+urllib3==1.26.4
+    # via
+    #   -r requirements/base.txt
     #   requests
     #   responses
 vine==1.3.0
     # via
-    #   -r base.txt
+    #   -r requirements/base.txt
     #   amqp
     #   celery
-wrapt==1.11.2
+wrapt==1.12.1
     # via astroid
