+++ requirements/tox.txt
@@ -12,7 +12,7 @@ filelock==3.0.12
     # via
     #   tox
     #   virtualenv
-packaging==20.4
+packaging==20.9
     # via tox
 pluggy==0.13.1
     # via tox
@@ -20,18 +20,17 @@ py==1.10.0
     # via tox
 pyparsing==2.4.7
     # via packaging
-six==1.15.0
+six==1.16.0
     # via
-    #   packaging
     #   tox
     #   virtualenv
 toml==0.10.2
     # via tox
 tox-battery==0.6.1
-    # via -r tox.in
-tox==3.20.1
+    # via -r requirements/tox.in
+tox==3.23.1
     # via
-    #   -r tox.in
+    #   -r requirements/tox.in
     #   tox-battery
-virtualenv==20.1.0
+virtualenv==20.4.6
     # via tox
